cleaning up thread
