+++ client/src/api/index.ts
@@ -7,6 +7,7 @@ export * from './queries/user';
 // Hook exports
 export * from './hooks/usePoll';
 export * from './hooks/useThreads';
+export * from './hooks/useThread';
 export * from './hooks/useCreatePost';
 export * from './hooks/useNotifications';
 export * from './hooks/useTopUsers';
