+++ client/src/api/queries/thread.ts
@@ -1,4 +1,4 @@
-import { ForumThread } from "@/lib/types";
+import { ForumThread, ThreadReply } from "@/lib/types";
 import { apiRequest } from "@/lib/queryClient";
 
 export const fetchPinnedThreads = async (category: string) => {
@@ -98,6 +98,189 @@ export const createThread = async (params: CreateThreadParams) => {
       throw errorData; // Throw the complete error object
     }
     throw new Error(errorData.message || "Failed to create post");
+  }
+
+  return response.json();
+};
+
+// Fetch single thread by ID
+export const fetchThreadById = async (threadId: string, userId?: string) => {
+  const response = await fetch(
+    `/api/threads/id/${threadId}${userId ? `?userId=${userId}` : ""}`
+  );
+  if (!response.ok) {
+    throw new Error(`Failed to fetch thread: ${response.statusText}`);
+  }
+  return response.json() as Promise<ForumThread>;
+};
+
+// Fetch thread replies
+export const fetchThreadReplies = async (threadId: string) => {
+  const response = await fetch(`/api/threads/${threadId}/replies`);
+  if (!response.ok) {
+    throw new Error(`Failed to fetch replies: ${response.statusText}`);
+  }
+  return response.json() as Promise<ThreadReply[]>;
+};
+
+// Like a thread
+export const likeThread = async (threadId: string, userId: string) => {
+  const response = await apiRequest(
+    "POST",
+    `/api/threads/${threadId}/like`,
+    { userId }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Dislike a thread
+export const dislikeThread = async (threadId: string, userId: string) => {
+  const response = await apiRequest(
+    "POST",
+    `/api/threads/${threadId}/dislike`,
+    { userId }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Toggle pin status for a thread by user
+export const toggleThreadPinByUser = async (threadId: string, userId: string) => {
+  const response = await apiRequest(
+    "POST", 
+    `/api/threads/${threadId}/pinned-by-user`,
+    { userId }
+  );
+  
+  if (!response.ok) {
+    throw new Error("Failed to pin post");
+  }
+  
+  return response.json();
+};
+
+// Submit a reply to a thread
+export const submitThreadReply = async (
+  threadId: string, 
+  userId: string, 
+  content: string,
+  parentReplyId: string | null
+) => {
+  const response = await apiRequest(
+    "POST",
+    `/api/threads/${threadId}/replies`,
+    {
+      userId,
+      content,
+      parentReplyId
+    }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+
+    // Check for the special upgrade required error
+    if (errorData.error === "UPGRADE_REQUIRED") {
+      throw new Error("UPGRADE_REQUIRED");
+    }
+
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Like a reply
+export const likeReply = async (replyId: number, userId: string) => {
+  const response = await apiRequest(
+    "POST",
+    `/api/replies/${replyId}/like`,
+    { userId }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Dislike a reply
+export const dislikeReply = async (replyId: number, userId: string) => {
+  const response = await apiRequest(
+    "POST",
+    `/api/replies/${replyId}/dislike`,
+    { userId }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Delete a thread
+export const deleteThread = async (threadId: string, userId: string, role?: string) => {
+  const response = await apiRequest(
+    "DELETE", 
+    `/api/threads/${threadId}`, 
+    { userId, role }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
+  }
+
+  return response.json();
+};
+
+// Delete a reply
+export const deleteReply = async (replyId: string, userId: string, role?: string) => {
+  const response = await apiRequest(
+    "DELETE", 
+    `/api/replies/${replyId}`, 
+    { userId, role }
+  );
+
+  if (!response.ok) {
+    const errorData = await response.json().catch(() => ({}));
+    throw new Error(
+      errorData.message ||
+        `Error: ${response.status} ${response.statusText}`
+    );
   }
 
   return response.json();
